requirements.txt
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
cssmin==0.2.0
Deprecated==1.2.18
Flask==3.1.0
Flask-Assets==2.1.0
Flask-Compress==1.17
Flask-Limiter==3.12
flask-talisman==1.1.0
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsmin==3.0.1
limits==5.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pycparser==2.22
Pygments==2.19.1
rich==13.9.4
sentry-sdk==2.27.0
typing_extensions==4.13.2
urllib3==2.4.0
webassets==2.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zstandard==0.23.0
Pillow==10.3.0
